--- Tests/Runtime/Utils.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: a7c2208cd8384da2abd0184aa374f2be
-timeCreated: 1698152230
