--- Tests/Runtime/Utils/ScreenshotHelperTest.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 9c559438a0f648d49f0c29fe4fc48a9a
-timeCreated: 1698152241
